requirements.txt
python-dotenv
requests
fastapi
uvicorn
pydantic
google-generativeai
beautifulsoup4
yfinance
matplotlib
pandas
